+++ CHANGELOG.md
@@ -142,6 +142,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2536 Gain some space in patient statistics widget
 - YLP-2539 Update arrow icon in Device parameters change history
 - YLP-2551 Sort device usage last settings by anti-chronological order
+- YLP-2554 Hide tooltip and values on TIR stat title when one bar is hovered
 - YLP-2557 Responsive for daily and trends views
 - YLP-2558 Statistics panel design improvements